Replace content index with store index everywhere (#146)
* WIP

* Fix test to use WriteConteIndex(To/From)Buffer

* Remove all references to content index in longtail.h

* Remove write of version content index

* Replace all uses of ContentIndex with StoreIndex

* const cast

* Fix naming

* Update cmd/main.c to use store index

* Use murmur3 for memstorage path hashes to avoid collisions

* Remove unused found_store_chunk_per_block in Longtail_GetExistingStoreIndex

* Change uint64_t -> uint32_t for block/chunk indexes

* proper cast for murmur32 hash function

* Make this release pre-release

* release notes

* Added Longtail_VersionIndex_GetChunkHashes Longtail_VersionIndex_GetChunkSizes Longtail_VersionIndex_GetChunkTags

* release notes

* Remove unneeded parameters from Longtail_GetExistingStoreIndex

* release notes

* release notes

* remove pre-relase flag

* update README.md regarding content index / store index

+++ lib/compressblockstore/longtail_compressblockstore.c
@@ -240,7 +240,7 @@ static int CompressBlockStore_PutStoredBlock(
 
 static int CompressBlockStore_PreflightGet(
     struct Longtail_BlockStoreAPI* block_store_api,
-    uint64_t chunk_count,
+    uint32_t chunk_count,
     const TLongtail_Hash* chunk_hashes)
 {
     MAKE_LOG_CONTEXT_FIELDS(ctx)
@@ -453,7 +453,7 @@ static int CompressBlockStore_GetStoredBlock(
 
 static int CompressBlockStore_GetExistingContent(
     struct Longtail_BlockStoreAPI* block_store_api,
-    uint64_t chunk_count,
+    uint32_t chunk_count,
     const TLongtail_Hash* chunk_hashes,
     uint32_t min_block_usage_percent,
     struct Longtail_AsyncGetExistingContentAPI* async_complete_api)
